Move component ConfirmedCasesByMunicipalitiesCard from 'components/cards' to 'components/index/CardsReference'
